fix: update total data calculation in useAccounts hook and adjust no_of_bills handling for improved accuracy

+++ util/account.js
@@ -32,7 +32,7 @@ export function getTotalsByStore(accounts) {
       };
     }
 
-    acc[item.store_id].no_of_bills += parseFloat(item.no_of_bills || 0);
+    acc[item.store_id].no_of_bills += parseInt(item.no_of_bills || 0);
     acc[item.store_id].total_sales += parseFloat(item.total_sales || 0);
     acc[item.store_id].card_sales += parseFloat(item.card_sales || 0);
     acc[item.store_id].sales_return += parseFloat(item.sales_return || 0);
